# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartQueue AI - Intelligent Virtual Queue Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /queues/{queueId}/tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List tokens in a queue
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
    post:
      operationId: joinQueue
      tags: [tokens]
      summary: Join a queue (get a token)
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinQueueBody"
      responses:
        "201":
          description: Joined queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
  /tokens/{tokenId}:
    get:
      operationId: getToken
      tags: [tokens]
      summary: Get a token by ID
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDetail"
  /tokens/{tokenId}/confirm:
    post:
      operationId: confirmArrival
      tags: [tokens]
      summary: Confirm arrival (QR scan)
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Arrival confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
  /tokens/{tokenId}/leave:
    post:
      operationId: leaveQueue
      tags: [tokens]
      summary: Leave the queue
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveQueueBody"
      responses:
        "200":
          description: Left the queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
  /tokens/{tokenId}/skip:
    post:
      operationId: skipToken
      tags: [tokens]
      summary: Skip a token (admin action)
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token skipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
  /tokens/{tokenId}/complete:
    post:
      operationId: completeToken
      tags: [tokens]
      summary: Mark token as served/completed
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, admin, staff]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]

    # Queue schemas
    Queue:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        industry:
          type: string
          enum: [hospital, bank, salon, college, government, retail, other]
        location:
          type: string
        avgServiceTimeMinutes:
          type: integer
        status:
          type: string
          enum: [active, paused, closed]
        maxCapacity:
          type: integer
        currentCount:
          type: integer
        estimatedWaitMinutes:
          type: integer
        openAt:
          type: string
        closeAt:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, industry, location, avgServiceTimeMinutes, status, currentCount, estimatedWaitMinutes, createdAt]
...
    Token:
      type: object
      properties:
        id:
          type: integer
        queueId:
          type: integer
        userId:
          type: integer
        tokenNumber:
          type: string
        position:
          type: integer
        status:
          type: string
          enum: [waiting, called, serving, completed, skipped, left, no_show]
        priority:
          type: string
          enum: [normal, vip, emergency]
        estimatedWaitMinutes:
          type: integer
        noShowProbability:
          type: number
        arrivalConfirmed:
          type: boolean
        joinedAt:
          type: string
          format: date-time
        calledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required: [id, queueId, userId, tokenNumber, position, status, priority, estimatedWaitMinutes, arrivalConfirmed, joinedAt]
    TokenDetail:
      allOf:
        - $ref: "#/components/schemas/Token"
        - type: object
          properties:
            queue:
              $ref: "#/components/schemas/Queue"
            user:
              $ref: "#/components/schemas/User"
            qrCode:
              type: string
            confidenceLevel:
              type: string
              enum: [high, medium, low]
    JoinQueueBody:
      type: object
      properties:
        priority:
          type: string
          enum: [normal, vip, emergency]
          default: normal
        notes:
          type: string
      required: []
    LeaveQueueBody:
      type: object
      properties:
        reason:
          type: string
      required: []

    # Analytics schemas
...